+++ src/SadConsole.Shared/GameHelpers/Scene.cs
@@ -11,259 +11,158 @@ using SadConsole.Input;
 namespace SadConsole.GameHelpers
 {
     /// <summary>
-    /// Groups a <see cref="LayeredSurface"/> and a list of <see cref="GameObject"/> types together.
+    /// Groups a <see cref="ISurface"/> and a list of <see cref="GameObject"/> types together.
     /// </summary>
-    [DataContract]
-    public class Scene: LayeredSurface
+    public class Scene: Screen
     {
         /// <summary>
         /// The objects for the scene.
         /// </summary>
-        [DataMember]
         public List<GameObject> Objects;
 
         /// <summary>
         /// Regions defined for the scene.
         /// </summary>
-        [DataMember]
         public List<Zone> Zones;
 
         /// <summary>
         /// Hotspots defined for the scene.
         /// </summary>
-        [DataMember]
         public List<Hotspot> Hotspots;
 
         /// <summary>
-        /// Width of the backing <see cref="LayeredSurface"/>.
+        /// Width of the backing <see cref="Surface"/>.
         /// </summary>
-        public int Width { get { return BackgroundSurface.Width; } }
+        public int Width { get { return Surface.Width; } }
 
         /// <summary>
-        /// Height of the backing <see cref="LayeredSurface"/>.
+        /// Height of the backing <see cref="Surface"/>.
         /// </summary>
-        public int Height { get { return BackgroundSurface.Height; } }
+        public int Height { get { return Surface.Height; } }
 
         /// <summary>
         /// Access to the backing console that is internally wrapped in this scene.
         /// </summary>
-        public Console BackingConsole { get { return baseConsole; } }
-
-        #region CONSOLE
-        public Point Position
-        {
-            get
-            {
-                return ((IConsole)baseConsole).Position;
-            }
-
-            set
-            {
-                ((IConsole)baseConsole).Position = value;
-                OnCalculateRenderPosition();
-            }
-        }
-
-        public ISurface TextSurface
-        {
-            get
-            {
-                return backgroundSurface;
-            }
-
-            set
-            {
-            }
-        }
-
-        public Cursor VirtualCursor
-        {
-            get
-            {
-                return ((IConsole)baseConsole).VirtualCursor;
-            }
-        }
-
-        public IScreen Parent
-        {
-            get
-            {
-                return ((IConsole)baseConsole).Parent;
-            }
-
-            set
-            {
-                ((IConsole)baseConsole).Parent = value;
-            }
-        }
-
-        public bool UsePixelPositioning
-        {
-            get
-            {
-                return ((IConsole)baseConsole).UsePixelPositioning;
-            }
-
-            set
-            {
-                ((IConsole)baseConsole).UsePixelPositioning = value;
-            }
-        }
-
-        public bool UseKeyboard
-        {
-            get
-            {
-                return ((IConsole)baseConsole).UseKeyboard;
-            }
-
-            set
-            {
-                ((IConsole)baseConsole).UseKeyboard = value;
-            }
-        }
-
-        public bool UseMouse
-        {
-            get
-            {
-                return ((IConsole)baseConsole).UseMouse;
-            }
-
-            set
-            {
-                ((IConsole)baseConsole).UseMouse = value;
-            }
-        }
-
-        public bool IsFocused
-        {
-            get
-            {
-                return ((IConsole)baseConsole).IsFocused;
-            }
-
-            set
-            {
-                ((IConsole)baseConsole).IsFocused = value;
-            }
-        }
-
-        public bool IsExclusiveMouse
-        {
-            get
-            {
-                return ((IConsole)baseConsole).IsExclusiveMouse;
-            }
-
-            set
-            {
-                ((IConsole)baseConsole).IsExclusiveMouse = value;
-            }
-        }
-
-        public bool IsVisible
-        {
-            get
-            {
-                return ((IConsole)baseConsole).IsVisible;
-            }
-
-            set
-            {
-                ((IConsole)baseConsole).IsVisible = value;
-            }
-        }
-
-        //public bool DoUpdate
-        //{
-        //    get
-        //    {
-        //        return ((IConsole)baseConsole).DoUpdate;
-        //    }
-
-        //    set
-        //    {
-        //        ((IConsole)baseConsole).DoUpdate = value;
-        //    }
-        //}
-
-        public bool ProcessMouse(MouseConsoleState state)
-        {
-            return ((IConsole)baseConsole).ProcessMouse(state);
-        }
-
-        public bool ProcessKeyboard(Keyboard info)
-        {
-            return ((IConsole)baseConsole).ProcessKeyboard(info);
-        }
-
-        public void Update(TimeSpan updateTime)
-        {
-            ((IConsole)baseConsole).Update(updateTime);
-        }
-
-        public void Draw(TimeSpan renderTime)
-        {
-            ((IConsole)baseConsole).Draw(renderTime);
-        }
-
-        public void LostMouse(MouseConsoleState state)
-        {
-            ((IConsole)baseConsole).LostMouse(state);
-        }
-        #endregion
+        public SurfaceEditor Surface { get; }
 
         /// <summary>
-        /// Called when the parent position changes.
+        /// The renderer used in drawing the <see cref="Surface"/>.
         /// </summary>
-        public virtual void OnCalculateRenderPosition()
-        {
-            relativePosition = Position;
-            IScreen parent = Parent;
+        public Renderers.ISurfaceRenderer SurfaceRenderer { get; set; }
 
-            while (parent != null)
-            {
-                relativePosition += parent.Position;
-                parent = parent.Parent;
-            }
-
-            foreach (var child in Children)
-            {
-                child.OnCalculateRenderPosition();
-            }
-        }
+        /// <summary>
+        /// Treats the <see cref="Position"/> of the console as if it is pixels and not cells.
+        /// </summary>
+        [DataMember]
+        public bool UsePixelPositioning { get; set; } = false;
 
         /// <summary>
         /// Creates a new Scene from an existing <see cref="LayeredSurface"/>.
         /// </summary>
         /// <param name="surface">The surface for the scene.</param>
-        public Scene(LayeredSurface surface)
+        /// <param name="renderer">The renderer for the surface.</param>
+        public Scene(ISurface surface, Renderers.ISurfaceRenderer renderer)
         {
-            baseConsole = new Console(surface);
-            backgroundSurface = surface;
+            Surface = new SurfaceEditor(surface);
+            SurfaceRenderer = renderer;
             Objects = new List<GameObject>();
             Zones = new List<Zone>();
             Hotspots = new List<Hotspot>();
         }
 
-        public static Scene Load(string file, Console baseConsole = null, params Type[] types)
+        /// <summary>
+        /// Creates a new surface using a <see cref="BasicSurface"/> at the specified width and height.
+        /// </summary>
+        /// <param name="width">The width of the surface.</param>
+        /// <param name="height">The height of the surface.</param>
+        public Scene(int width, int height): this(new BasicSurface(width, height), new Renderers.SurfaceRenderer())
         {
-            var scene = SadConsole.Serializer.Load<Scene>(file, types);
 
-            if (baseConsole == null)
-                scene.baseConsole = new Console(scene.backgroundSurface) { Renderer = new Renderers.LayeredSurfaceRenderer() };
-            else
-                scene.baseConsole = baseConsole;
+        }
+
+        public override void Draw(TimeSpan timeElapsed)
+        {
+            if (IsVisible)
+            {
+                // Draw the scene.
+                SurfaceRenderer.Render(Surface.TextSurface);
+                Global.DrawCalls.Add(new DrawCallSurface(Surface.TextSurface, calculatedPosition, UsePixelPositioning));
+
+                // Draw objects
+                foreach (var item in Objects)
+                    item.Draw(timeElapsed);
+
+                // Draw the children.
+                if (Children.Count != 0)
+                {
+                    var copyList = new List<IScreen>(Children);
+
+                    foreach (var child in copyList)
+                        child.Draw(timeElapsed);
+                }
+            }
+        }
+
+        Point cachedOffset;
+
+        public override void Update(TimeSpan timeElapsed)
+        {
+            base.Update(timeElapsed);
+            
+            Point offset = position + Surface.TextSurface.RenderArea.Location;
+
+            if (offset != cachedOffset)
+            {
+                cachedOffset = offset;
+
+                foreach (var gameObject in Objects)
+                    gameObject.PositionOffset = cachedOffset;
+            }
+        }
+
+        public override void OnCalculateRenderPosition()
+        {
+            base.OnCalculateRenderPosition();
+
+            Point offset = position + Surface.TextSurface.RenderArea.Location;
+
+            if (offset != cachedOffset)
+            {
+                cachedOffset = offset;
+
+                foreach (var gameObject in Objects)
+                    gameObject.PositionOffset = cachedOffset;
+            }
+        }
+
+        /// <summary>
+        /// Saves the scene to a file. You must serialize the <see cref="Surface"/> separately.
+        /// </summary>
+        /// <param name="file">The destination file.</param>
+        public void Save(string file)
+        {
+            Serializer.Save((SerializedTypes.SceneSerialized)this, file);
+        }
+
+        /// <summary>
+        /// Loads scene from a file.
+        /// </summary>
+        /// <param name="file">The source file.</param>
+        /// <param name="surface">The surface for the scene.</param>
+        /// <param name="renderer">The renderer used to draw the surface.</param>
+        /// <returns></returns>
+        public static Scene Load(string file, ISurface surface, Renderers.ISurfaceRenderer renderer)
+        {
+            var scene = (Scene)Serializer.Load<SerializedTypes.SceneSerialized>(file);
+
+            scene.Surface.TextSurface = surface;
+            scene.SurfaceRenderer = renderer;
+
+            scene.OnCalculateRenderPosition();
 
             return scene;
         }
 
-        public void Save(string file, params Type[] types)
-        {
-            SadConsole.Serializer.Save<Scene>(this, file, types);
-        }
 
-        
     }
 }